Fix Double to Long cast

+++ src/main/java/org/lockss/laaws/rs/io/index/solr/SolrArtifactIndex.java
@@ -758,7 +758,7 @@ public class SolrArtifactIndex implements ArtifactIndex {
             QueryResponse response = solr.query(q);
             FieldStatsInfo contentLengthStats = response.getFieldStatsInfo().get("contentLength");
 
-            return (Long)contentLengthStats.getSum();
+            return ((Double)contentLengthStats.getSum()).longValue();
         } catch (SolrServerException e) {
             throw new IOException(e);
         }
